Add several options to project creation dialog
Specifically, users can now set project info, including:
- who the authors are
- what the title is
- and whether it's an empirical or formal theory project (or both)
- (FIXME: get input from qual scholars on what they'd need)

They can also specify a bib file to copy to their new project directory,
and can select a few options regarding git:
- whether to initiate a git repo for the new project (default yes)
- whether to ignore PDF output in .gitignore (default yes)
- whether to ignore binary data files in .gitignore (default yes)

Finally, they can specify whether to have a separate 'analysis'
subdirectory to break some of their code out into (a) separate file(s).
Generally this should be discouraged, but it can be appropriate in some
settings (this is discussed in the Rmd file copied to the 'analysis'
subdirectory if it is created).